Revert "feat: Decimal item/armor weight value (#4)"
This reverts commit d7c4ea82785177459d7f6f416941a3ea761e951a.

+++ src/Battlescape/AIModule.cpp
@@ -6488,7 +6488,7 @@ float AIModule::getItemPickUpScore(BattleItem* item)
 	if (!valid)
 		return 0;
 	score = item->getRules()->getSellCost();
-	float mainHandWeight = 0;
+	int mainHandWeight = 0;
 	if (_unit->getMainHandWeapon())
 		mainHandWeight = _unit->getMainHandWeapon()->getRules()->getWeight();
 	float encumbrance = (float)_unit->getBaseStats()->strength / (float)(_unit->getCarriedWeight() - mainHandWeight + item->getRules()->getWeight());
